Refactor filings, move into institutions file
